requirements.txt
google-generativeai==0.3.1
python-dotenv==0.19.0